lots of bugs in the old setfullexpression() files. totally redone, just one function for all binary functions now, and hopefully all bugs fixed too. still not 100% sure, so try and break it if you have time.
NOTE the book says that AND comes BEFORE OR in terms of preference, not equal to each other. As such, i have altered the code to respect that. if we ask Neff and he says do otherwise, it is easily changeable.

Also added functionality for IFF. no button yet, since it isn't required yet. the symbol is <-> or just < if you want to use it.